[fix] CurrentWeatherResult 모델 파일로 분리
분리
